redid project for masters class, added testing

+++ ModelUtils.cs
@@ -8,11 +8,11 @@ namespace xModel_Gen
 {
     public static class ModelUtils
     {
-        public static int GetDistance(double x1, double y1, double x2, double y2)
+        public static double GetDistance(double x1, double y1, double x2, double y2)
         {
-            return (int)Math.Sqrt(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2));
+            return Math.Sqrt(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2));
         }
-        public static int GetDistance(Node first, Node second)
+        public static double GetDistance(Node first, Node second)
         {
             return GetDistance(first.GridX, first.GridY, second.GridX, second.GridY);
         }
